+++ source/extensions/omni.isaac.orbit/test/sim/test_spawn_from_files.py
@@ -31,6 +31,8 @@ class TestSpawningFromFiles(unittest.TestCase):
 
     def setUp(self) -> None:
         """Create a blank new stage for each test."""
+        # Create a new stage
+        stage_utils.create_new_stage()
         # Simulation time-step
         self.dt = 0.1
         # Load kit helper
@@ -45,6 +47,8 @@ class TestSpawningFromFiles(unittest.TestCase):
         # stop simulation
         self.sim.stop()
         self.sim.clear()
+        self.sim.clear_all_callbacks()
+        self.sim.clear_instance()
 
     """
     Basic spawning.