+++ source/extensions/omni.isaac.orbit/test/sim/test_spawn_materials.py
@@ -32,6 +32,8 @@ class TestSpawningMaterials(unittest.TestCase):
 
     def setUp(self) -> None:
         """Create a blank new stage for each test."""
+        # Create a new stage
+        stage_utils.create_new_stage()
         # Simulation time-step
         self.dt = 0.1
         # Load kit helper
@@ -44,6 +46,8 @@ class TestSpawningMaterials(unittest.TestCase):
         # stop simulation
         self.sim.stop()
         self.sim.clear()
+        self.sim.clear_all_callbacks()
+        self.sim.clear_instance()
 
     def test_spawn_preview_surface(self):
         """Test spawning preview surface."""